requirements.txt
textual
rich
pygments
google-genai
